refactor(index): split monolith into smaller modules
